add Universe and update cells based on neighbours
We're currently considering a static window size. Upon resizing, the grid should stay the same but the cell size should adjust

For now all the Cells are inititialized Dead so the changes won't be noticeable
We currently use a vector to store the grid, but a matrix might be better in the future

The typing in the update function should be reviewed and tidied up.