Add asynchronous request responding
This change was necessary since previously a thread was spawned for each request.
Now a thread pool only spawns only 4 threads, no matter how many requests arrive.

The task management works as follows: A ThreadPool will create a channel and hold on to the sending side of the channel. Each Worker will hold on to the receiving side of the channel. Every Job will hold the closures we want to send down the channel. The ThreadPool will send the job it wants to execute down the sending side of the channel. Finally, the Worker will loop over its receiving side of the channel and execute the closures of any jobs it receives.

Until now, I haven't found a use for the crossbeam crate. But now, there is a performance issue that can be solved with some datastructure and thread synchronization tools offered by crossbeam.