Add threadpool abstraction
We added the thread_pool.rs script as lib. In order to do that we needed to add another lib manifest to the Cargo.toml. The problem with this is that It is currently the case that every cargo package corresponds to exactly 0 or 1 library crates. Its not possible to use cargo to create a package containing multiple libraries.
So to get this done, we created a file (lib.rs) that contains both the http & thread_pool module.

Another problem that appeared with this is that every server with each feature needed to compile both modules in order to build succesfuly. But, the tokio & async implementations needed to have crossbeam as a dependency because of this. So we added the #[cfg(feature="<feature>")] macro that only requires the compilation of that module in a specific feature