Http server crossbeam implementation (#10)
* Add cargo new files

* Add TCP Listener that accepts the connections

* Add request read handling

* Add basic request response

Hardcoded code 200 response to any kind of request, with no body nor headers

* Add multithreading connections

* Add get root response html & not found response

Hardcoded, for now...

* Remove unnecesary use of move

* Add asynchronous request responding

This change was necessary since previously a thread was spawned for each request.
Now a thread pool only spawns only 4 threads, no matter how many requests arrive.

The task management works as follows: A ThreadPool will create a channel and hold on to the sending side of the channel. Each Worker will hold on to the receiving side of the channel. Every Job will hold the closures we want to send down the channel. The ThreadPool will send the job it wants to execute down the sending side of the channel. Finally, the Worker will loop over its receiving side of the channel and execute the closures of any jobs it receives.

Until now, I haven't found a use for the crossbeam crate. But now, there is a performance issue that can be solved with some datastructure and thread synchronization tools offered by crossbeam.

* Add Makefile for building & running the server

* Add server shutdown implementation

Drop implementation for the ThreadPool

* Add POST handling

* Add cargo clippy command to build directive

* Fix number of requests accepted

It was limited because I was testing manually that the workers terminated & then shutted down

* Add build-release directive

* Disabled clippy linter in the handle_read function

This is because clippy recommended to use read_exact in the handle_read function but there are cases where the buffer is not filled and then the server panics with this message: thread '<unnamed>' panicked at 'called `Result::unwrap()` on an `Err` value: Error { kind: UnexpectedEof, message: "failed to fill whole buffer" }'

* Refactor response creation

* Add crossbeam crate to dependencies

* Replaced mpsc channel with crossbeam unbounded

* Remove prints & runned cargo fmt

* Removed worker's id

* Refactor names & abstractions

* Moved server directory to inner directory

* Refactor code structure

- Removed unnecesary files
- Moved main.rs from crossbeam src to crossbeam in order to use it as a lib
- Untracked templates

* Add config for crossbeam server implementation

* Add integration with http abstraction

* Remove useless println

* Run cargo fmt

* Removed unused templates

* Add threadpool abstraction

We added the thread_pool.rs script as lib. In order to do that we needed to add another lib manifest to the Cargo.toml. The problem with this is that It is currently the case that every cargo package corresponds to exactly 0 or 1 library crates. Its not possible to use cargo to create a package containing multiple libraries.
So to get this done, we created a file (lib.rs) that contains both the http & thread_pool module.

Another problem that appeared with this is that every server with each feature needed to compile both modules in order to build succesfuly. But, the tokio & async implementations needed to have crossbeam as a dependency because of this. So we added the #[cfg(feature="<feature>")] macro that only requires the compilation of that module in a specific feature

* Fix unused directory