refactor TcpStream Read/Write
stream.write_all is used instead of stream.write. BufReader and
BufWriter are not used. As read from the documentation:

"BufWriter<W> can improve the speed of programs that make small and
repeated write calls to the same file or network socket. It does not
help when writing very large amounts at once, or writing just one or a
few times. It also provides no advantage when writing to a destination
that is in memory, like a Vec<u8>."